# internal/api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Enduro API
  version: "1.0"
servers:
- url: http://localhost:9000
paths:
...
  /pipeline:
    get:
      tags:
      - pipeline
      summary: list pipeline
      description: List all known pipelines
      operationId: pipeline#list
      parameters:
      - name: name
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: Sunt est quae est.
        example: Veniam aut qui blanditiis provident.
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnduroStoredPipeline'
                example:
                - id: 95545D09-3153-0F17-09F7-7FEB1545D780
                  name: Doloribus ut amet.
                - id: 95545D09-3153-0F17-09F7-7FEB1545D780
                  name: Doloribus ut amet.
                - id: 95545D09-3153-0F17-09F7-7FEB1545D780
                  name: Doloribus ut amet.
                - id: 95545D09-3153-0F17-09F7-7FEB1545D780
                  name: Doloribus ut amet.
              example:
              - id: 95545D09-3153-0F17-09F7-7FEB1545D780
                name: Doloribus ut amet.
              - id: 95545D09-3153-0F17-09F7-7FEB1545D780
                name: Doloribus ut amet.
  /pipeline/{id}:
    get:
      tags:
      - pipeline
      summary: show pipeline
      description: Show pipeline by ID
      operationId: pipeline#show
      parameters:
      - name: id
        in: path
        description: Identifier of pipeline to show
        required: true
        schema:
          type: string
          description: Identifier of pipeline to show
          example: 1950F5A3-DD57-36ED-FC7A-02434F6D102C
          format: uuid
        example: 17130190-1B4F-4068-0538-3D42B488D22B
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnduroStoredPipeline'
              example:
                id: 6E989AEF-9C11-4862-D8B1-5CFC4F6414E0
                name: Dolor animi aspernatur sed assumenda ea.
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 9292983799343016043
                message: Voluptatibus consequatur in quaerat dolorum.
...
components:
  schemas:
...
    EnduroStoredPipeline:
      type: object
      properties:
        id:
          type: string
          description: Name of the collection
          example: 751556FB-5B96-05F5-4890-014C4C315530
          format: uuid
        name:
          type: string
          description: Name of the collection
          example: A quia quia quo.
      description: StoredPipeline describes a pipeline retrieved by this service.
      example:
        id: AFE55386-F5B6-467C-35B0-2117A9683AF1
        name: Voluptas autem aut quis reiciendis corporis.
      required:
      - name
...
    NotFound:
      type: object
      properties:
        id:
          type: integer
          description: Identifier of missing collection
          example: 2135324167707328023
        message:
          type: string
          description: Message of error
          example: Et autem adipisci et occaecati quaerat.
      description: Collection not found
      example:
        id: 11276450690720251243
        message: Ab pariatur.
      required:
      - message
      - id
...